+++ shared/node.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/pokt-network/pocket/consensus"
-	"github.com/pokt-network/pocket/libp2p"
 	"github.com/pokt-network/pocket/logger"
 	"github.com/pokt-network/pocket/p2p"
 	"github.com/pokt-network/pocket/persistence"
@@ -39,13 +38,6 @@ func CreateNode(bus modules.Bus, options ...modules.ModuleOption) (modules.Modul
 }
 
 func (m *Node) Create(bus modules.Bus, options ...modules.ModuleOption) (modules.Module, error) {
-	// TECHDEBT: simplify after P2P module consolidation.
-	useLibP2P := bus.GetRuntimeMgr().GetConfig().UseLibP2P
-	p2pCreate := p2p.Create
-	if useLibP2P {
-		p2pCreate = libp2p.Create
-	}
-
 	for _, mod := range []func(modules.Bus, ...modules.ModuleOption) (modules.Module, error){
 		state_machine.Create,
 		persistence.Create,
@@ -54,7 +46,7 @@ func (m *Node) Create(bus modules.Bus, options ...modules.ModuleOption) (modules
 		telemetry.Create,
 		logger.Create,
 		rpc.Create,
-		p2pCreate,
+		p2p.Create,
 	} {
 		if _, err := mod(bus); err != nil {
 			return nil, err
@@ -165,6 +157,11 @@ func (m *Node) GetBus() modules.Bus {
 // TODO: Move all message types this is dependant on to the `messaging` package
 func (node *Node) handleEvent(message *messaging.PocketEnvelope) error {
 	contentType := message.GetContentType()
+	logger.Global.Debug().Fields(map[string]any{
+		"message":     message,
+		"contentType": contentType,
+	}).Msg("node handling event")
+
 	switch contentType {
 	case messaging.NodeStartedEventType:
 		logger.Global.Info().Msg("Received NodeStartedEvent")